fix: minor code quality improvements
